Changed inside the iscolliding method - do diffaltitude returns a double

+++ ATMHandin3/ATMHandin3/Classes/CollisionAvoidanceSystem.cs
@@ -63,7 +63,7 @@ namespace ATMHandin3.Classes
 
         private bool IsColliding(Aircraft ac1, Aircraft ac2)
         {
-            int diffAltitude = Calculate.CalculateAltitudeDiff(ac1.Altitude, ac2.Altitude);
+            double diffAltitude = Calculate.CalculateAltitudeDiff(ac1.Altitude, ac2.Altitude);
             double diffLongitude = Calculate.DistanceTo(ac1.XCoordinate, ac2.XCoordinate, ac1.YCoordinate, ac2.YCoordinate);
             return diffAltitude <= _altitudeTolerance && diffLongitude <= _longitudeTolerance;
         }
